requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
langchain-groq
faiss-cpu
sentence-transformers
python-dotenv
